Finish implementing digamma as an internal command.  Size is about 300 bytes, 210 for xrom code, 80 for constants and one command table entry plus catalogues.
The series expansion has been extended by a dozen terms which seems
to produce 20 or more digits accuracy.  A conditional compile allows
another ten terms to be incuded that seem to buy another 7-8 digits
for a space cost of almost 170 more bytes.

The CNST and [cmplx]CNST commands are enabled since they are being
used.  The space estimates here don't include the overheads for
this command.

A further space saving measure can be performed by making the real code
call the complex directly.  This will be slower but accuracy shouldn't suffer.



git-svn-id: https://svn.code.sf.net/p/wp34s/code@2510 9654ca62-1852-461c-aa49-1be32f3bb02d